historian/writer/requirements.txt
confluent-kafka==2.4.0
influxdb-client==1.44.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
